Week 6 Part 3

+++ app/Http/Controllers/PostController.php
@@ -4,10 +4,20 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['index', 'show']),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -54,7 +64,10 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        //
+        // Authorizing the action
+        Gate::authorize('modify', $post);
+
+
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -63,7 +76,10 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        //
+        // Authorizing the action
+        Gate::authorize('modify', $post);
+
+        // validate
         $fields = $request -> validate([
             'title' =>['required', 'max:255'],
             'body' =>['required']
@@ -77,6 +93,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        // Authorizing the action
+        Gate::authorize('modify', $post);
+
         //
         $post ->delete();
         return back()->with('delete', 'your post was deleted');
